finalize symlinking
